chore(i10): reorganize mod structure
