+++ src/weather_aggregator/weather_clients.rs
