+++ src/weather_aggregator/weather_clients/open_weather.rs
